platform/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
supabase
pymongo
openai
langchain
langgraph
